+++ src/main/java/com/mcm/sp/dto/CategoriaDTO.java
@@ -11,7 +11,7 @@ import com.mcm.sp.entities.Categoria;
 public class CategoriaDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
-	private Long id;
+	private Integer id;
 	
 	@NotEmpty(message="Preenchimento obrigatório!!")
 	@Length(min=5, max=80, message="O tamanho deve ser entre 5 e 80 caracteres!!")
@@ -24,16 +24,16 @@ public class CategoriaDTO implements Serializable {
 		nome = categoria.getNome();
 	}
 	
-	public CategoriaDTO(Long id, String nome) {
+	public CategoriaDTO(Integer id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
